requirements.txt
pandas>=2.1.0
numpy>=1.24.0
yfinance>=0.2.31
ta>=0.11.0
alpaca-trade-api>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
plotly>=5.18.0
dash>=2.14.0
python-dotenv>=1.0.0
schedule>=1.2.0
loguru>=0.7.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0
scikit-learn>=1.3.0
pyarrow>=14.0.0
feedparser>=6.0.0
